+++ src/main/java/com/hytsnbr/shiny_colors/ShinyColorsCdInfoCollectorApplication.java
@@ -1,0 +1,18 @@
+package com.hytsnbr.shiny_colors;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+
+@SpringBootApplication
+@ConfigurationPropertiesScan
+public class ShinyColorsCdInfoCollectorApplication {
+    
+    public static void main(String[] args) {
+        System.exit(
+            SpringApplication.exit(
+                SpringApplication.run(ShinyColorsCdInfoCollectorApplication.class, args)
+            )
+        );
+    }
+}
