+++ src/main/java/com/hytsnbr/shiny_colors/dto/BaseJsonData.java
@@ -1,0 +1,21 @@
+package com.hytsnbr.shiny_colors.dto;
+
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/** JSON関連基底クラス */
+public abstract class BaseJsonData {
+    
+    /** ファイル作成日（エポック日） */
+    @JsonProperty("createdAt")
+    protected final long createdAt = LocalDateTime
+        .now()
+        .atZone(ZoneId.systemDefault())
+        .toEpochSecond();
+    
+    public long getCreatedAt() {
+        return this.createdAt;
+    }
+}
