--- src/main/java/com/hytsnbr/shiny_test/ShinyColorsCdInfoCollectorApplication.java
@@ -1,18 +1,0 @@
-package com.hytsnbr.shiny_test;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-
-@SpringBootApplication
-@ConfigurationPropertiesScan
-public class ShinyColorsCdInfoCollectorApplication {
-    
-    public static void main(String[] args) {
-        System.exit(
-            SpringApplication.exit(
-                SpringApplication.run(ShinyColorsCdInfoCollectorApplication.class, args)
-            )
-        );
-    }
-}
