--- src/main/java/com/hytsnbr/shiny_test/exception/SystemException.java
@@ -1,21 +1,0 @@
-package com.hytsnbr.shiny_test.exception;
-
-import org.apache.commons.lang3.StringUtils;
-
-/** システムエラー例外クラス */
-public class SystemException extends RuntimeException {
-    
-    private static final String MSG = "システムエラー";
-    
-    public SystemException(String msg) {
-        this(msg, null);
-    }
-    
-    public SystemException(Exception exception) {
-        this(null, exception);
-    }
-    
-    public SystemException(String msg, Exception exception) {
-        super(StringUtils.isNotBlank(msg) ? msg : MSG, exception);
-    }
-}
